fix(tempfile): Add option to set the name of temporary files
Since I have realized that we will need to copy all of the temporary files into the project folder, we really should be using nicer file names instead of uuids.  For local runs, I think this will also cut down on the amount of "garbage tempfile" space that we start to take up on people's machines, since the file will get overwritten each time a model with the same ID or a wea from the same city is written.