fix(daylight): update handlers to use full_id
This commit adds support for models that are created with sensor-grid-groups. In that case, the results can be in a sub-folder with the group name. grid_info now provides a `full_id` key which is `identifier/group_identifier` for objects with group and stays as `identifier` for grids with no group assigned.

The change will not break the older models as it checks to see if `full_id` key is available before trying to get the value.