fix(model): Revert #136 grid check with files
We used to have the Model handler set up such that, if you just connected up the file path to a HBJSON, the file would just get copied for upload without loading it into Python. [It seems that we changed this a few months ago](https://github.com/pollination/handlers-python/commit/62fe930f5bc76a99250f9fdd8824ae3fb755260f) but we should revert it since being able to plugin in a file path instead of a Model object and know that extra time wouldn't be spent saving or re-parsing it was really useful.

From my understanding, the re-parsing routine was added a few months ago because of checks that we wanted to run from the Rhino plugin but, for a case like that, the check should be run on the .NET side where we have the Rhino plugin Model object in .NET form instead of writing a large .NET Model object to JSON, loading the model from JSON to Python, running the check in Python, and then writing the Model from Python back to JSON. [It's for cases like this that we have handlers that are written in .NET.](https://github.com/pollination/handlers-csharp)